Refactor Mocks in completion tests
  - Move common Mocks into a separate Mocks class so that both Stream/TaskCompletionProviderTests can make use of it

Resolves #1743
